Move FillStyle
