+++ src/CanvasAPI/FillStyle.js
@@ -1,7 +1,7 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as CanvasPattern$WebAPI from "../CanvasAPI/CanvasPattern.js";
-import * as CanvasGradient$WebAPI from "../CanvasAPI/CanvasGradient.js";
+import * as CanvasPattern$WebAPI from "./CanvasPattern.js";
+import * as CanvasGradient$WebAPI from "./CanvasGradient.js";
 
 function decode(t) {
   if (CanvasGradient$WebAPI.isInstanceOf(t)) {
