+++ src/DOMAPI.res
@@ -198,8 +198,6 @@ type shareData = {
   mutable url?: string,
 }
 
-@editor.completeFrom(FillStyle) type fillStyle
-
 /**
 @editor.completeFrom(Window) The location (URL) of the object it is linked to. Changes done on it are reflected on the object it relates to. Both the Document and Window interface have such a linked Location, accessible via Document.location and Window.location respectively.
 [See Location on MDN](https://developer.mozilla.org/docs/Web/API/Location)
@@ -9517,7 +9515,7 @@ type canvasRenderingContext2D = {
   /**
     [Read more on MDN](https://developer.mozilla.org/docs/Web/API/CanvasRenderingContext2D/strokeStyle)
     */
-  mutable strokeStyle: unknown,
+  mutable strokeStyle: fillStyle,
   /**
     [Read more on MDN](https://developer.mozilla.org/docs/Web/API/CanvasRenderingContext2D/fillStyle)
     */
