+++ tests/DOMAPI/HTMLCanvasElement__test.js
@@ -1,6 +1,6 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as FillStyle$WebAPI from "../../src/DOMAPI/FillStyle.js";
+import * as FillStyle$WebAPI from "../../src/CanvasAPI/FillStyle.js";
 
 let myCanvas = document.getElementById("myCanvas");
